+++ docs/changelog.md
@@ -13,6 +13,7 @@ permalink: /changelog/
 
 * **feature-top-sites**
   * ⚠️ **This is a breaking change**: This changes `fetchProvidedTopSites` in `TopSitesConfig` into a data class `TopSitesProviderConfig` that specifies whether or not to display the top sites from the provider. [#11654](https://github.com/mozilla-mobile/android-components/issues/11654)
+  * ⚠️ **This is a breaking change**: Removed `cachedTopSites` in `DefaultTopSitesStorage`. [#11691](https://github.com/mozilla-mobile/android-components/issues/11691)
 
 # 98.0.0
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v97.0.0...v98.0.0)